Refactor: getPostCount() 메서드 반환 타입을 int로 변경

+++ src/main/java/com/ebstudy/board_v2/repository/BoardMapper.java
@@ -11,7 +11,7 @@ public interface BoardMapper {
 
     List<PostDTO> getPostList(int pageNumber);
 
-    long getPostCount();
+    int getPostCount();
 
     PostDTO getPost(long postId);
